Refactor texture, spritesheet and font to use VAO in rendering
Using one single VAO to render multiple of instance of the same resource
(KRR_TEXTURE, KRR_SPRITE, or KRR_FONT). As well optimized to modify
vertex data only when necessary as in the case of KRR_TEXTURE with
clipping thus in this case use glBufferSubData() only when need, but use
the original pre-create data from start if nothing change.

KRR_FONT uses the same VAO as created in KRR_SPRITE as the former relies
on the functionality of the latter; build and operate on top.

As well; this leaves for later optimization, we can possibly delete VAO
of KRR_TEXTURE inside KRR_SPRITE after vertex buffer generation phase.
Because KRR_SPRITE doesn't use any rendering functionality of
KRR_TEXTURE; instead it has its own vertex data and indices buffer under
management.

All in all, in initialization phase of user code, make sure to always
create shader for texture and font then set it shared shaders first so
all of those 3 struct will be able to create and function without crash.