Fix fixed-timestep. Update window to expose focus gained/lost event. Immprove performance when app not active.
When the app is not active, now it will reduces the framerate to 10 FPS
while at the same time use SDL_Delay() to sleep with 1/10 s. Sluttering
becomes more apparent, thus resolved this issue for now with enabling of
vsync in main.c.

Fixed timestep is fixed to not always call render if not yet reach the
time. Both update() and render() will be supplied with common_frameTime
which is the delta time since last frame after it reaches the time to
call such update, not anymore a fixed delta time as each loop it won't
be exact, and update/render call will probably catch up the target set
framerate anyway.