Add headless sample
The way this works is to supply -DSDL_HEADLESS at compile time, then in main.c it will detect and initialize properly according to our request either with or without window. As all sample code shared the same main.c to initially common and higher level stuff, thus it's like it is in this changes.