Add terrain (rendering/generation) support + sample code
Initially duplicated the functionality that SIMPLEMODEL, and
texturedpp3d shaderprogram has but with slightly more added
functionality and for future that it will be definitely diverge in terms
of functionalities to support.

The sample code doesn't have movement in space support just yet, so we
need to rotate the terrain along the x-axis, but you have to keep
pressing W to go up in y-axis to see it.

terrain.c sample modified to support projection matrix for both UI and
3D view (using different matrix for each one). Perspective is more
correct in terms of rendering for 3D view (no distorted as perceived in
orthographic one). So we stick to it at the moment.