Improve (and correction) textured3d and terrain3d shader
This fixed specular lighting calculation we have in shader code. We had
accessed individual position information of camera in wrong way. For
safety, just multiply with vec4(0.0, 0.0, 0.0, 1.0) in order to get the
right answer.

Possibly this has to do with col-major storage that opengl uses
internally. Thus what we see in layout of view matrix in row-major, we need to
transpose to get col-major in which it matches to the result of what we
tested trying to get an individual element via inverted_mat[3][0] to get
position-x, and inverted_mat[3][1] to get position-y, and so on.