Maintain 2 version of matrices for UI and 3D stuff
We might probably need this whenever our screen resolution changed, thus aspect ratio changed. At that time, we might not need to scale our 3d model in which case if we do, we need to scale light's position, might be others too. Thus we add model matrix for UI stuff.