Fix and improve objloader
Fix the major bug for .obj file that doesn't have duplicated vertex thus it won't set the number of result vertex correctly. Now this change calculates that number properly.

As well as improved the checking statement to grow the output vertices memory space.

Note: The model that doesn't have duplicated vertex is grassModel.obj
which leads to this fix.