Fix and majorly improve timestep
The root cause to investigate thus leads to this changes is due to start
up time which make diff time so large that calculated diff time is wrong
significantly and it cannot bring the value down. Because of delta time
each frame is not large enough than accumulated time.

Read https://gafferongames.com/post/fix_your_timestep/ and apply such
solution into this by introducing catching up loop to consume multiple
of unit (stable) delta time for number of times if necessary for each
frame. The remainder will be carried on for the next frame. You can try
to count for number of time it executes inside the while() catching up
loop, then print the value out after such loop. It will go like
1,1,1,1,1,1,...,2,1,1,1,1,1,....2. As 2 pops up, this is to consume all
the reminders that piled up.

This changes stabilize timestep of game loop!