Add graphic util function to generate mipmaps-stack after loading texture
The approach to use this function is to firstly load the target texture.
Then call glBindTexture() to bind such texture, then call this util
function to generate mipmap.

Doing like this to keep the memory consumption as low as possible as the
engine won't assume generating mipmap for all texture unnecessary.

As well, this include setting of level of detail bias value. More on negative
value side will further nullity the effect of mipmap, but more positive
value will make it stronger.