Add collision detection (y-pos) for player in terrain sample & fix updating jump velocity
We introduce a slight offset when we need to compare to ground level
to change from in-air state to normal (on ground state) for player. This
will have effect of bouncing back a little bit as lerping will work this
out normally in update_camera() function.

Fixed updating player's jump velocity for setting initial jump power. We
just need to do it once. Now it's correct.