Update terrain sample to rotate objects placing on terrain properly
The technique uses barycentric function to carry out the operation. The
function compute_terrain_normal() accepts 3d position. Its
implementation is to use KRR_math_barycentric_xz() function to get the
result of a single value in which we use this function swapping
neighbor's y-position value for x, y, and z component of normal vector
in order for us to interpolate normal result.