Fix major bug of main.c for initialize all sample programs
It's ALWAYS and a MUST to define 'static' in front of the function signature if we mean to use it only in that source file.
If not, and we link with other various libraries, there is a chance that other libraries will make a function call to our
same name of function without us aware. See https://stackoverflow.com/questions/31304591/odd-segmentation-fault-with-cc-gcc-but-not-g-c-sdl2-linux#comment97489592_31304591
for direct related problem.