Add preprocessing header for detecting platforms
It's not immediate in use at the moment, but might be used in the future.
It will set KRR_TARGET_PLATFORM preprocessor variable as included by common.h for both foundation and graphics at the top,
then we can do platform specific code by checking if such platform preprocessor variable is defined.

As well, it will set preprocessor variable for mobile and PC category; just in case we don't need to know which specific platform is it.