Fix to supply headless compile flag to headless sample instead of function call
setenv() or putenv() seems to be platform specific which is risky to be
using as you might as well have to define a certain macro first before
use in which as tested it doesn't work. So we avoid it entirely and use
this supplying via compile flag intead for reliability.