Fix texture locking
Before the code is made for OpenGL, now it is migrated and fixed for OpenGL ES 3.0.
To read pixels from texture directly in OpenGL ES, it's not possible as we need to bind and read it through
framebuffer. So this created a new framebuffer name, bind it, read pixels via glReadPixels then unbind and delete
framebuffer when done.